% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.Density.R
\name{as.Density}
\alias{as.Density}
\alias{as.Density.default}
\alias{as.Density.Density}
\alias{as.Density.density}
\alias{as.Density.sm}
\alias{as.Density.ssden}
\alias{as.Density.kde}
\alias{as.Density.logspline}
\alias{as.Density.npdensity}
\alias{as.Density.pendensity}
\alias{as.Density.ash}
\title{Density Conversion}
\usage{
as.Density(object, ...)

\method{as.Density}{default}(object, xin, ...)

\method{as.Density}{Density}(object, ...)

\method{as.Density}{density}(object, xin, ...)

\method{as.Density}{sm}(object, xin, ...)

\method{as.Density}{ssden}(object, n = 301, ...)

\method{as.Density}{kde}(object, xnam, ...)

\method{as.Density}{logspline}(object, xin, augment = TRUE, n = 301, ...)

\method{as.Density}{npdensity}(object, n = 301, ...)

\method{as.Density}{pendensity}(object, n = 301, augment = TRUE, ...)

\method{as.Density}{ash}(object, xin, smooth = TRUE, n = 301, ...)
}
\arguments{
\item{object}{the output from one of the selected kernel density methods.}

\item{\dots}{optional additional arguments required by certain methods.}

\item{xin}{the input data for the selected kernel density method (required).}

\item{n}{the number of points to compute if \code{smooth} is \code{TRUE}.}

\item{xnam}{the name of the variable.}

\item{augment}{logical, if \code{TRUE}, then augment the truncated ends of
a spline density estimate to make the area as close to 1 as possible. If \code{FALSE},
then do not adjust the density estimate.}

\item{smooth}{logical, if \code{TRUE}, then smooth the original estimate. If
\code{FALSE}, then do not smooth the original estimate.}
}
\value{
An object of class "Density" with \code{object} included in the
object component to facilitate exploring the original object.
}
\description{
Functions to convert selected kernel density objects to objects of class "Density."
}
\note{
The default method assumes at least a list with components \code{x},
which contains the ordinates that are uniformly distributed and \code{y},
which are the density estimates. The integral of y by x must be close to 1.\cr
The default method also traps objects created by \code{sm} that have a specific
structure, but have no class that ties them to that function.\cr
The method for ash objects extrapolates the ends to a zero density value based
on the slope at the endpoints if "ash estimate nonzero outside interval ab."
Otherwise, the density at adjacent points is set to 0.
}
