% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootDensity.R
\name{bootDensity}
\alias{bootDensity}
\title{Bootstrap Resampling}
\usage{
bootDensity(x, method, bootN = 200, CI = 0.9, ...)
}
\arguments{
\item{x}{the numeric variable for which density estimates are needed.}

\item{method}{character the name of the density function to use. See 
\bold{Details}.}

\item{bootN}{the number of bootstrap iterations. The default is 200.}

\item{CI}{the confidence interval to estimate. The default is 0.90.}

\item{\dots}{additional arguments for \code{method}. The arguments must be fully
specified, abbreviations are not allowed.}
}
\value{
An object of class "Density" with the bootstrap analysis in the 
components \code{boot} and \code{bootCI}.
}
\description{
Generates bootstrap analysis of selected density estimates. The mean and 
confidence intervals are computed.
}
\details{
\code{Method} must be one of "qqDensity," "dsDensity," "density,"
"ash," "ssden," "local," "logspline," "bootdata," or "merge." In
general, \code{method} is the name of the function, but "ash" indicates 
\code{ash1} and "local" indicates \code{density.lf}. For specialized density
estimates, "bootdata" resamples the data so that consistent bootstrap analyses
can be computed and "merge" merges the bootstrap estimates into a Density object.
An example is provided in the Overview vignette.\cr
The only option for kernel density estimation is "density." The results of the 
various methods are nearly identical and the bandwidth can be set from one of 
the other methods.\cr
For \code{method} = "merge," \code{x} must be the output matrix of the resampled
density estimates and the Density object must be supplied in an argument called
\code{Density}.\cr
The \code{mixDensity} and \code{pendensity} methods are not implemented 
because the fit cannot be specified to ensure that the method assumes that the
resampled data come from the same population as the original. The same kind of 
issues can also affect \code{logDensity}, but the user can set up to ensure
that assumption. See the example in the Overview vignette.
}
\note{
In general, \code{set.seed} should be set before calling 
\code{bootDensity} to ensure reproducibility and consistency across different
functions.
}
