% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.os.R
\name{bw.os}
\alias{bw.os}
\alias{bw.lf}
\title{Bandwidth for Kernel Density Estimation}
\usage{
bw.os(x, kernel = "normal")

bw.lf(x)
}
\arguments{
\item{x}{numeric vector of observations from the distribution whose density 
is to be estimated.}

\item{kernel}{character string indicating the type of kernel. Must be one of
"normal," "box," "epanech," "biweight," or "triweight." The default is
"normal." See \bold{Details}.}
}
\value{
A single numeric value representing the value.
}
\description{
Computes the bandwidth from the variance of x, specifically the oversmoothed 
bandwidth selector of Wand and Jones (1995, page 61) for \code{bw.os} or
the modified bandwidth for \code{density.lf} (\code{bw.lf}).
}
\details{
The value for \code{kernel} is required for functions that do not
automatically adjust the bandwidth for the variance of the kernel, like 
\code{bkde}. The function \code{density}, for example, does and the kernel
should always be "normal."
}
