% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenSamp.R
\docType{data}
\name{DenSamp}
\alias{DenSamp}
\title{Sample Data}
\format{
Vector of 100 samples
}
\usage{
DenSamp
}
\description{
Arbitrary sample data to demonstrate some features of density estimates.
}
\examples{
data(DenSamp)
head(DenSamp)
}
\keyword{datasets}
