% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density_dist.R
\name{dDensity}
\alias{dDensity}
\alias{Density_dist}
\alias{pDensity}
\alias{qDensity}
\title{Density Distribution}
\usage{
dDensity(x, object)

pDensity(q, object)

qDensity(p, object)
}
\arguments{
\item{x}{vector of quantities.}

\item{object}{an object of class "Density."}

\item{q}{vector of quantities.}

\item{p}{vector of probabilities.}
}
\value{
The function \code{ddensity} gives the density, \code{pdensity} gives
the cumulative distribution, and \code{qdensity} gives the quantiles of the
empirical density estimate. One value is returned for each element of the first
argument.
}
\description{
Density, cumulative distribution, and quantile functions for objects of
class "Density."
}
