% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsDensity.R
\name{dsDensity}
\alias{dsDensity}
\alias{dsDensity.default}
\alias{dsDensity.Surv}
\title{Double Smoothing Density}
\usage{
dsDensity(x, h, n = 301, from, to)

\method{dsDensity}{default}(x, h, n = 301, from, to)

\method{dsDensity}{Surv}(x, h, n = 301, from, to)
}
\arguments{
\item{x}{the variable for which density estimates are needed.}

\item{h}{the bandwidth, if missing, then Silverman’s rule of thumb
is used. See \code{Details}.}

\item{n}{the number of output density points.}

\item{from, to}{the left and right limits of the range of desired output points
where the density is to be estimated. If missing, then reasonable limits are
computed beyond the range of the data.}
}
\value{
A list of class "Density" with \code{n} values of x, the ordinate; 
and y the density values; summary statistics and tuning parameters; and the call.
}
\description{
An implementation of the double-smoothing density described by Yao.
}
\details{
Silverman’s rule of thumb, computes the bandwidth as 0.9 times the 
minimum of the standard deviation and the interquartile range divided by 1.34 
times the sample size to the negative one-fifth power. (From the documentation
for \code{bw.nrd0} in the stats package.) This tends to work better for double
smoothing than the normal density estimate, because it can handle smaller values
of bandwidth than traditional methods.
}
