% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofScores.R
\name{gofScores}
\alias{gofScores}
\title{Goodness of Fit Scores}
\usage{
gofScores(x, ..., plot.scores = FALSE)
}
\arguments{
\item{x}{the original sample data}

\item{\dots}{any number of objects of class "Density." To create abbreviations
for the Density objects, they can be named in the call.}

\item{plot.scores}{logical, if \code{TRUE}, then plot the ZaNorm and LogLik scores 
with a reference line for the kernel density estimate relation. If \code{FALSE},
the default, then no plot is created. The points are labelled if the Density
objects are named in the call.}
}
\value{
a data frame of the ZaNorm, VarD, and VarRat scores, and the log-likelihood
(LogLik) and degrees of freedom (DF) with one row per 
density object. The data are sorted by ZaNorm.
}
\description{
Computes the Zhang modified goodness of fit test type A, without computing the 
p-level. The value is normalized such that 0 indicates a perfect fit (ZaNorm). A 
convolution measure of the variability of the density estimate is 
also computed as VarDen. VarRat is the ratio of the variance of the fit to that 
of the data. Used primarily to asses the relative goodness of fits between 
several density estimates.
}
\details{
The convolution measure of the variability of the density estimate
(VarDen) is computed as the sum of the squared differences between 101 uniformly 
distributed points in the range of \code{x} (100 differences). The square-root
of the value divided by the peak value of density and is adjusted so that a
normal distribution gives a value of 1.0.
}
\note{
Smaller values of Za indicate closer matches to the distribution.
}
