% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDensity.R
\name{logDensity}
\alias{logDensity}
\alias{logDensity.default}
\alias{logDensity.Surv}
\title{Density Estimate for Log-Transformed Data}
\usage{
logDensity(x, method, n = 301, ...)

\method{logDensity}{default}(x, method, n = 301, ...)

\method{logDensity}{Surv}(x, method, n = 301, ...)
}
\arguments{
\item{x}{the variable for which density estimates are needed.}

\item{method}{the name of the density estimation method to use on the 
log-transformed data. The variable for which density estimates are needed 
must be \code{x}.}

\item{n}{the number of output density points for the-back-transformed values.}

\item{\dots}{any number of additional arguments for \code{method}}
}
\value{
An object of class "Density."
}
\description{
Some highly skewed data or data that have a minimum of 0 benefit from a 
log-transformation and possible probability adjustment for 0 values.
}
\details{
Zero values are extracted from \code{x} and the remaining values are
log-transformed. The back-transformed probability distribution is adjusted for
the proportion of zero values.
}
