% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixDensity.R
\name{mixDensity}
\alias{mixDensity}
\title{Mixed Gaussian Density Estimate}
\usage{
mixDensity(x, Nmix = 2, n = 301, ...)
}
\arguments{
\item{x}{the variable for which density estimates are needed.}

\item{Nmix}{the number of Gaussian mixtures in \code{x}. If 1, then the density
is computed assuming the data are normally distributed. The default is 2.}

\item{n}{the number of output density points.}

\item{\dots}{Additional arguments for specific methods. Not used currently.}
}
\value{
A list of class "Density" with \code{n} values of x, the ordinate; 
and y the density values; stats, the input and derived (computed) summary 
statistics; h the value of h used in the estimate; and call, the call. In 
addition to the standard components of a Density object, \code{posterir},
the posterior probabilities; \code{logLike}, the log likelihood; and 
\code{mix}, the mean, standard deviation, and \code{posterior} the proportion 
of each Gaussian distribution for every point.
}
\description{
A density estimation method that moels the data as a mixture of two or more
Gaussian distributions. This method does not support censored data.
}
\details{
The value of \code{alpha} can affect the peakedness of the
density estimate---small values, near 0, have higher peaks than those 
near 0.5, which produces wider tails. The default is 0.5\cr
The argument \code{k} determines the number of knots in the cubic spline fit.
The minimum is 4 larger the values can decrease the smoothness of the fit.
The default is 5, which appears to be a good compromise.\cr
The default \code{qqDensity} method selects a distribution based on the 
statistics of \code{x}. If the skewness is greater than 0.5, then a gamma 
distribution is selected. If the skewness is less than -0.5, then a flipped 
gamma distribution is selected. If the kurtosis is greater than 3.67, then a 
t distribution is selected. If the kurtosis is less than 2.5, then a 
generalized normal distribution is selected. If the statistics of \code{x} 
meets non of those criteria, then a normal distribution is used. The Surv
method always uses a logistic distribution.
}
