% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Density.R
\name{plot.Density}
\alias{plot.Density}
\title{Plot a Density Object}
\usage{
\method{plot}{Density}(x, xlab, ylab, type = "l", ...)
}
\arguments{
\item{x}{object of class "Density."}

\item{xlab}{the x-axis label. If missing, then the label is the variable name.}

\item{ylab}{the y-axis label. If missing, then the label is automatically 
generated from information in \code{x}.}

\item{type}{the type of plot, default is "l," which draws the fit as a line.}

\item{\dots}{additional arguments modifying the plot.}
}
\value{
The object is returned invisibly.
}
\description{
Plot the computed density and points for a density object.
}
