% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Density.R
\name{print.Density}
\alias{print.Density}
\title{Print Density Object}
\usage{
\method{print}{Density}(x, digits = 4, ...)
}
\arguments{
\item{x}{the Density object.}

\item{digits}{the number of significant digits to print in the numeric values.}

\item{\dots}{further arguments for other methods.}
}
\description{
Print a general summary of a Density object.
}
