% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqDensity.R
\name{qqDensity}
\alias{qqDensity}
\alias{qqDensity.default}
\alias{qqDensity.Surv}
\title{Q-Q Density Estimate}
\usage{
qqDensity(x, alpha = 0.5, k = 5, n = 301, from, to, ...)

\method{qqDensity}{default}(
  x,
  alpha = 0.5,
  k = 5,
  n = 301,
  from,
  to,
  shape = "auto",
  plot.fit = TRUE,
  ...
)

\method{qqDensity}{Surv}(x, alpha = 0.5, k = 5, n = 301, from, to, plot.fit = TRUE, ...)
}
\arguments{
\item{x}{the variable for which density estimates are needed.}

\item{alpha}{the plotting position tuning parameter. See \bold{Details}.}

\item{k}{the number of knots in the monotonic cubic spline.  See \bold{Details}.}

\item{n}{the number of output density points.}

\item{from, to}{the left and right limits of the range of desired output points
where the density is to be estimated. If missing, then reasonable limits are
computed beyond the range of the data.}

\item{\dots}{Additional arguments for specific methods.}

\item{shape}{the general description of the shape of the distribution of 
\code{x}. Must be one of "auto" (default), "normal," "left," "right," "platy,"
"lepto," or "multimodal." See \bold{Details}.}

\item{plot.fit}{logical controlling a plot of the Q-Q fit. If \code{TRUE}, then
the plot is created on the current plotting device. Otherwise, no plot is 
created.}
}
\value{
A list of class "Density" with \code{n} values of x, the ordinate; 
and y the density values; stats, the input and derived (computed) summary 
statistics; h the value of h used in the estimate; and call, the call.
}
\description{
A proposed density estimate computed by fitting a smooth, monotonically 
increasing curve through the quantiles of a distribution related to 
the observed values to compute a cumulative probability estimate (CDF). The 
density estimate is the derivative of the CDF. The purpose is to better 
describe the observed data.
}
\details{
The value of \code{alpha} can affect the tails of the
density estimate---small values, near 0, have more narrow tails than those 
near 0.5, which produces wider tails. The default is 0.5\cr
The argument \code{k} determines the number of knots in the cubic spline fit.
The minimum is 4 larger the values can decrease the smoothness of the fit.
The default is 5, which appears to be a good compromise.\cr
The default \code{shape} method selects a distribution based on the 
statistics of \code{x}. If the skewness is greater than 0.5, then a gamma 
distribution is selected, specified by "right." If the skewness is less than 
-0.5, then a flipped gamma distribution is selected, specified by "left."
If the kurtosis is greater than 3.67, then a t distribution is selected,
specified by "lepto." If the kurtosis is less than 2.5, then a generalized 
normal distribution is selected, specified by "platy." If the statistics of \code{x} 
meets none of those criteria, then a normal distribution is used, specified by
"normal." If the shape is multimodal, then the user must specify "multimodal."
Can be abbreviated. The Surv method always uses a normal distribution.
}
