% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosEst.R
\name{rosEst}
\alias{rosEst}
\title{Censored Estimation}
\usage{
rosEst(Obs, Cens, dens, alpha = 0.4)
}
\arguments{
\item{Obs}{the reported values. Missing values are not permitted.}

\item{Cens}{the left-censor codes, if \code{TRUE}, then the corresponding
value in \code{Obs} is left-censored, otherwise the corresponding value is
the detected value. Missing values are not permitted.}

\item{dens}{a list containing x (ordinate) and y (density) values 
describing the distribution.}

\item{alpha}{the plotting position offset factor, default is 0.4.}
}
\value{
A vector like \code{Obs} with the left-censored values replaced with
the estimated values.
}
\description{
Internal function  to compute the regression on order statistics (ROS)-
estimation values for censored values.
}
\note{
ROS-substitution works on single- or multiple-detection limit data.
}
